backend/ltml: Subdivide main Parser module
