+++ backend/src/Language/Ltml/Parser/Paragraph.hs
@@ -8,7 +8,8 @@ import Control.Monad.State (evalStateT)
 import Language.Lsd.AST.Type.Paragraph (ParagraphType (ParagraphType))
 import Language.Ltml.AST.Node (Node (Node))
 import Language.Ltml.AST.Paragraph (Paragraph (Paragraph))
-import Language.Ltml.Parser (Parser, nonIndented)
+import Language.Ltml.Parser (Parser)
+import Language.Ltml.Parser.Common.Indent (nonIndented)
 import Language.Ltml.Parser.Common.Lexeme (nLexeme1)
 import Language.Ltml.Parser.Label (labelingP)
 import Language.Ltml.Parser.Text (ParagraphParser, textForestP)
