+++ backend/src/Language/Ltml/Parser/Section.hs
@@ -19,7 +19,8 @@ import Language.Ltml.AST.Section
     , Section (Section)
     , SectionBody (..)
     )
-import Language.Ltml.Parser (Parser, nonIndented, wrapParser)
+import Language.Ltml.Parser (Parser, wrapParser)
+import Language.Ltml.Parser.Common.Indent (nonIndented)
 import Language.Ltml.Parser.Common.Lexeme (nLexeme)
 import Language.Ltml.Parser.Footnote (FootnoteParser)
 import Language.Ltml.Parser.Footnote.Combinators (manyWithFootnotesTillSucc)
